refactoring place card
