+++ src/components/places/near-place-card/near-place-card.tsx
@@ -1,0 +1,17 @@
+import {ReactElement} from 'react';
+import {Offer} from '../../../types/offer.ts';
+import PlaceCard from '../place-card/place-card.tsx';
+
+type NearPlaceCardProps = {
+  offer: Offer;
+}
+
+function NearPlaceCard({offer}: NearPlaceCardProps): ReactElement {
+  return (
+    <article className="near-places__card place-card">
+      <PlaceCard offer={offer}/>
+    </article>
+  );
+}
+
+export default NearPlaceCard;
