+++ src/components/places/near-places/near-places.tsx
