+++ src/components/places/place-card-list/place-card-list.tsx
@@ -1,6 +1,6 @@
 import {Offer} from '../../../types/offer.ts';
-import PlaceCard from '../place-card/place-card.tsx';
 import {useState} from 'react';
+import CitiesPlaceCard from '../cities-place-card/cities-place-card.tsx';
 
 type PlaceCardListProps = {
   offers: Offer[];
@@ -15,7 +15,7 @@ function PlaceCardList({offers}: PlaceCardListProps) {
 
   return (
     <div className="cities__places-list places__list tabs__content">
-      {offers.map((offer) => <PlaceCard key={offer.id} offer={offer} mouseOverHandler={mouseOverHandler}/>)}
+      {offers.map((offer) => <CitiesPlaceCard key={offer.id} offer={offer} mouseOverHandler={mouseOverHandler}/>)}
     </div>
   );
 }
